split domain out into module of files
- we also added the validator crate for emails here
- and the fake crate for testing emails
- and the quickcheck crate for broader tests
