+++ src/domain/mod.rs
@@ -1,0 +1,9 @@
+mod subscriber_name;
+mod subscriber_email;
+mod new_subscriber;
+mod string_checks;
+
+pub use subscriber_name::SubscriberName;
+pub use subscriber_email::SubscriberEmail;
+pub use new_subscriber::NewSubscriber;
+pub use string_checks::*;
